refactor: user package
유저와 관련있는 부분들 모두 user 패키지로 이동
