+++ src/main/java/io/zerogone/controller/api/BlogUpdateController.java
@@ -2,7 +2,7 @@ package io.zerogone.controller.api;
 
 import io.zerogone.model.dto.BlogDto;
 import io.zerogone.model.dto.BlogMemberDto;
-import io.zerogone.model.dto.UserDto;
+import io.zerogone.user.model.UserDto;
 import io.zerogone.model.entity.MemberRole;
 import io.zerogone.service.BlogUpdateService;
 import io.zerogone.service.fileupload.ImageUploadService;
