+++ src/main/java/io/zerogone/controller/api/CommentCreateController.java
@@ -2,7 +2,7 @@ package io.zerogone.controller.api;
 
 import io.zerogone.model.dto.CommentDto;
 import io.zerogone.model.dto.PostDto;
-import io.zerogone.model.dto.UserDto;
+import io.zerogone.user.model.UserDto;
 import io.zerogone.service.create.CreateService;
 import io.zerogone.service.search.SearchService;
 import org.springframework.http.HttpStatus;
