+++ src/main/java/io/zerogone/controller/api/PostCreateController.java
@@ -3,7 +3,7 @@ package io.zerogone.controller.api;
 import io.zerogone.exception.NotAuthorizedException;
 import io.zerogone.model.dto.BlogDto;
 import io.zerogone.model.dto.PostDto;
-import io.zerogone.model.dto.UserDto;
+import io.zerogone.user.model.UserDto;
 import io.zerogone.service.create.CreateService;
 import io.zerogone.service.search.SearchService;
 import io.zerogone.validator.NewEntity;
