+++ src/main/java/io/zerogone/converter/CommentDtoToEntityConverter.java
@@ -2,10 +2,10 @@ package io.zerogone.converter;
 
 import io.zerogone.model.dto.CommentDto;
 import io.zerogone.model.dto.PostDto;
-import io.zerogone.model.dto.UserDto;
+import io.zerogone.user.model.UserDto;
 import io.zerogone.model.entity.Comment;
 import io.zerogone.model.entity.Post;
-import io.zerogone.model.entity.User;
+import io.zerogone.user.model.User;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
