+++ src/main/java/io/zerogone/service/BlogInvitationAcceptenceService.java
@@ -1,7 +1,7 @@
 package io.zerogone.service;
 
 import io.zerogone.model.dto.BlogDto;
-import io.zerogone.model.dto.UserDto;
+import io.zerogone.user.model.UserDto;
 import io.zerogone.model.entity.Blog;
 import io.zerogone.model.entity.BlogMember;
 import io.zerogone.repository.BlogDao;
