+++ src/main/java/io/zerogone/service/create/BlogCreateService.java
@@ -5,9 +5,9 @@ import io.zerogone.model.dto.BlogMemberDto;
 import io.zerogone.model.entity.Blog;
 import io.zerogone.model.entity.BlogMember;
 import io.zerogone.model.entity.MemberRole;
-import io.zerogone.model.entity.User;
+import io.zerogone.user.model.User;
 import io.zerogone.repository.BlogDao;
-import io.zerogone.repository.UserDao;
+import io.zerogone.user.UserDao;
 import io.zerogone.service.BlogInvitationService;
 import io.zerogone.service.fileupload.ImageUploadService;
 import io.zerogone.service.fileupload.ImageUrl;
