+++ src/main/java/io/zerogone/user/controller/LoginController.java
@@ -1,8 +1,8 @@
-package io.zerogone.controller.api;
+package io.zerogone.user.controller;
 
 import io.zerogone.exception.NotExistDataException;
-import io.zerogone.model.Email;
-import io.zerogone.model.dto.UserDto;
+import io.zerogone.user.model.Email;
+import io.zerogone.user.model.UserDto;
 import io.zerogone.service.search.SearchService;
 import io.zerogone.validator.Login;
 import org.springframework.validation.annotation.Validated;
