+++ src/main/java/io/zerogone/user/controller/LogoutController.java
@@ -1,4 +1,4 @@
-package io.zerogone.controller.api;
+package io.zerogone.user.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
