+++ src/main/java/io/zerogone/user/controller/SignupController.java
@@ -1,4 +1,4 @@
-package io.zerogone.controller;
+package io.zerogone.user.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
