+++ src/main/java/io/zerogone/user/converter/UserEntityToDtoConverter.java
@@ -1,7 +1,7 @@
-package io.zerogone.converter;
+package io.zerogone.user.converter;
 
-import io.zerogone.model.dto.UserDto;
-import io.zerogone.model.entity.User;
+import io.zerogone.user.model.UserDto;
+import io.zerogone.user.model.User;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
