+++ src/main/java/io/zerogone/user/model/Email.java
@@ -1,4 +1,4 @@
-package io.zerogone.model;
+package io.zerogone.user.model;
 
 import javax.validation.constraints.NotBlank;
