+++ src/main/java/io/zerogone/user/model/User.java
@@ -1,5 +1,6 @@
-package io.zerogone.model.entity;
+package io.zerogone.user.model;
 
+import io.zerogone.model.entity.Blog;
 import org.hibernate.annotations.Fetch;
 import org.hibernate.annotations.FetchMode;
 import org.hibernate.annotations.WhereJoinTable;
